# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  tags: string[];
  author: string;
  authorRole: string;
  /** @nullable */
  authorAvatar?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  readingTime: number;
  publishedAt: string;
  createdAt: string;
}

export interface ProcessStep {
  step: number;
  title: string;
  description: string;
}

export interface Faq {
  question: string;
  answer: string;
}

export interface Service {
  id: number;
  slug: string;
  title: string;
  shortDescription: string;
  fullDescription: string;
  category: string;
  icon: string;
  benefits: string[];
  process: ProcessStep[];
  technologies: string[];
  faqs: Faq[];
}

export interface Metric {
  label: string;
  value: string;
  /** @nullable */
  change?: string | null;
}

export interface CaseStudy {
  id: number;
  slug: string;
  clientName: string;
  industry: string;
  service: string;
  challenge: string;
  solution: string;
  results: Metric[];
  /** @nullable */
  featuredImage?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  createdAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  company: string;
  content: string;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  service: string;
}

export interface PortfolioItem {
  id: number;
  slug: string;
  title: string;
  client: string;
  category: string;
  description: string;
  /** @nullable */
  imageUrl?: string | null;
  technologies: string[];
  /** @nullable */
  projectUrl?: string | null;
  completedAt: string;
}

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  service?: string | null;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  budget?: string | null;
}

export interface NewsletterInput {
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface SubmitResult {
  success: boolean;
  message: string;
}

export interface AgencyStats {
  clientsServed: number;
  projectsCompleted: number;
  yearsExperience: number;
  countriesReached: number;
  averageRoiIncrease: number;
  googleRating: number;
  totalReviews: number;
}

export interface FeaturedContent {
  blogPosts: BlogPost[];
  caseStudies: CaseStudy[];
  testimonials: Testimonial[];
}

export type ListBlogPostsParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
limit?: number | null;
/**
 * @nullable
 */
offset?: number | null;
};

export type ListCaseStudiesParams = {
/**
 * @nullable
 */
service?: string | null;
};

export type ListPortfolioItemsParams = {
/**
 * @nullable
 */
category?: string | null;
};