# lib/api-zod/src/generated/types/agencyStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */

export interface AgencyStats {
  clientsServed: number;
  projectsCompleted: number;
  yearsExperience: number;
  countriesReached: number;
  averageRoiIncrease: number;
  googleRating: number;
  totalReviews: number;
}