# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  content: string;
  category: string;
  tags: string[];
  author: string;
  authorRole: string;
  /** @nullable */
  authorAvatar?: string | null;
  /** @nullable */
  featuredImage?: string | null;
  readingTime: number;
  publishedAt: string;
  createdAt: string;
}