# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  service?: string | null;
  /** @minLength 10 */
  message: string;
  /** @nullable */
  budget?: string | null;
}