# lib/api-zod/src/generated/types/metric.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */

export interface Metric {
  label: string;
  value: string;
  /** @nullable */
  change?: string | null;
}