# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */
import type { Faq } from './faq';
import type { ProcessStep } from './processStep';

export interface Service {
  id: number;
  slug: string;
  title: string;
  shortDescription: string;
  fullDescription: string;
  category: string;
  icon: string;
  benefits: string[];
  process: ProcessStep[];
  technologies: string[];
  faqs: Faq[];
}