# lib/api-zod/src/generated/types/submitResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SEO Agency Website API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResult {
  success: boolean;
  message: string;
}